# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/entree": {
            "post": {
                "description": "Create a new entree and return the new record's data to the caller",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entrees"
                ],
                "summary": "create entree",
                "parameters": [
                    {
                        "description": "The input entree data (only `option_name` is required)",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Entree"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an entree and returns a response to indicate success or failure",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entrees"
                ],
                "summary": "deletes an entree",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Entree ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    }
                }
            }
        },
        "/entrees": {
            "get": {
                "description": "Gets the selected entree for the given user ID (empty array if no selection has been made), or a list of all available entrees if no user ID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entrees"
                ],
                "summary": "gets one or all entrees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Logs in a user and returns the user details for the user (if authentication is successful)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logs in a user",
                "parameters": [
                    {
                        "description": "Log in details",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Signs up a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signs up a new user",
                "parameters": [
                    {
                        "description": "Sign up details",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserSignupInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/entrees": {
            "get": {
                "description": "Gets the selected entree for the given user ID (empty array if no selection has been made), or a list of all available entrees if no user ID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entrees"
                ],
                "summary": "gets one or all entrees",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_ENTREE"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Entree": {
            "type": "object",
            "required": [
                "option_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "option_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.HorsDoeuvres": {
            "type": "object",
            "required": [
                "option_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "option_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "description": "The user's email (must be unique); this field is an index.",
                    "type": "string"
                },
                "entreeSelection": {
                    "$ref": "#/definitions/models.Entree"
                },
                "entree_selection_id": {
                    "description": "The ID of the entree the user has selected; is null until the user makes a selection.",
                    "type": "string"
                },
                "first_name": {
                    "description": "The user's first name.",
                    "type": "string"
                },
                "horsDoeuvresSelection": {
                    "$ref": "#/definitions/models.HorsDoeuvres"
                },
                "hors_doeuvres_selection_id": {
                    "description": "The ID of the hors doeuvres the user has selected; is null until the user makes a selection.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_going": {
                    "description": "Whether or not the user is attending.",
                    "type": "boolean"
                },
                "last_name": {
                    "description": "The user's last name.",
                    "type": "string"
                },
                "password": {
                    "description": "Either a hash of the user's password (when stored in the DB), or a plain-text representation of the password (plain-text version is never stored)",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "The user's auth refresh token.",
                    "type": "string"
                },
                "role": {
                    "description": "The user's role, which can be \"GUEST\", \"INVITEE\" or \"ADMIN\". Defaults to \"GUEST\".",
                    "type": "string"
                },
                "token": {
                    "description": "The user's auth token.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.EntreeData": {
            "type": "object",
            "properties": {
                "entrees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Entree"
                    }
                }
            }
        },
...
        "types.UserData": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
...
        "types.UserLoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "types.UserSignupInput": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "types.V1_API_RESPONSE_ENTREE": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.EntreeData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.V1_API_RESPONSE_USERS": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.UserData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}